Fix remaining byte parameter in the API for steps.
Maybe this was the reason for #91 (negative steps)

This could happen making at least 3 steps per second.

+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/database/DBHandler.java
@@ -24,7 +24,7 @@ public interface DBHandler {
 
     List<ActivitySample> getSleepSamples(int tsFrom, int tsTo, SampleProvider provider);
 
-    void addGBActivitySample(int timestamp, byte provider, short intensity, byte steps, byte kind);
+    void addGBActivitySample(int timestamp, byte provider, short intensity, short steps, byte kind);
 
     SQLiteDatabase getWritableDatabase();
 }
